refactor: Convert AppBar component to TypeScript
